requirements.txt
google-genai>=0.4.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
streamlit>=1.32.0 